requirements.txt
openai-whisper>=20230314
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
anthropic>=0.39.0
piper-tts>=1.2.0
requests>=2.31.0
